+++ CHANGELOG.md
@@ -12,3 +12,10 @@ on the official Visual Studio extension gallery.
 **2016-07-05**
 
 - [x] Possibility to deploy the builded file to a configured folder
+
+## 1.0.0.1
+
+**2016-07-07**
+
+- [x] Read properties also from C++ projects
+- [x] Create an output pane for result logging